+++ git-packetline/CHANGELOG.md
@@ -9,3 +9,8 @@
     - `immutable::Band` -> `BandRef`
     - `immutable::DecodeBandError` -> `decode::band::Error`
     - `pub immutable::` -> `line::`
+    - `pub write::` -> `write::`
+
+* **removals**
+   - `write::Writer` (is now only `Writer`)
+   - `read::StreamingPeekableIter` (is now only `StreamingPeekableIter`)